pyproject.toml
[project]
name = "dokabun"
version = "0.1.0"
description = "Spreadsheet-driven CLI to populate LLM outputs. / スプシをインターフェースとして、LLMにどかっと分析させるツール"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Nu424" },
]
keywords = ["llm", "openrouter", "spreadsheet", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]
requires-python = ">=3.12"
dependencies = [
    "openai>=2.8.1",
    "httpx>=0.27.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "scikit-learn",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
# UMAP-based dimensionality reduction for `eo*` post-processing (e.g. `eou2`).
umap = [
    "umap-learn>=0.5.11",
]

[project.urls]
Repository = "https://github.com/Nu424/dokabun"
Issues = "https://github.com/Nu424/dokabunissues"

[project.scripts]
dokabun = "dokabun.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dokabun"]
